pyproject.toml
[tool.poetry]
name = "python-django-exercise"
version = "0.1.0"
description = ""
authors = ["Fabio Prizio <fabio.prizio@casavo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
black = "^24.4.2"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
